+++ release/_release.sql
@@ -1,291 +1,143 @@
+PRO Installing 3.0.0 (Kscope19)
 
--- =============================================================================
+--  sqlblanklines - Allows for SQL statements to have blank lines
+set sqlblanklines on
+--  define - Sets the character used to prefix substitution variables
+set define '^'
 
-insert into ks_session_status (display_seq, code, name, active_ind) values (35, 'ALT', 'Alternate', 'Y');
 
 
+-- *** DDL ***
 
--- =============================================================================
+-- #17
+alter table ks_events add blind_vote_flag varchar2(1);
+comment on column ks_events.blind_vote_begin_date is 'begin date of Public voting';
+comment on column ks_events.blind_vote_end_date is 'end date of Public voting';
+comment on column ks_events.blind_vote_flag is 'Indicates that the Public Voting will be "blind"';
 
--- install new version of the staging table to accomodate the new columns
-@../install/ks_full_session_load.sql
+-- #11
+alter table ks_sessions add room_size_code varchar2(20);
+comment on column ks_sessions.room_size_code is 'Define the size for a room S|M|L';
 
+-- #13
+alter table ks_users add expired_passwd_flag varchar2(1);
+alter table ks_users add login_attempts number;
+alter table ks_users add last_login_date date;
+comment on column ks_users.expired_passwd_flag is 'Set to Y when the account password is expired.';
+comment on column ks_users.login_attempts is 'Number of unsuccessful login attempts since last login';
+comment on column ks_users.last_login_date is 'Date the user was las successful login in';
 
--- install staging table's "lookup table" for what columns in it match what columns in the export file
-@../install/ks_load_mapping.sql
-@../conversion/seed_ks_load_mapping.sql
 
+-- #16
+-- Changes made by Ben Shumway
+alter table ks_full_session_load add presented_anything_ind varchar2(4000);
+alter table ks_full_session_load add presented_anything_where varchar2(4000);
+comment on column ks_full_session_load.presented_anything_ind is 'Whether the presenter has ever done a live presentation, anywhere for anything.';
+comment on column ks_full_session_load.presented_anything_where is 'Where the presenter has done live presentations (of any kind).';
 
--- new tables to accomdate roles
-@../install/ks_roles.sql
-@../conversion/seed_ks_roles.sql
+alter table ks_sessions add presented_anything_ind varchar2(1);
+alter table ks_sessions add presented_anything_where varchar2(4000);
+alter table ks_sessions add constraint ks_sessions_pres_any_yn check (presented_anything_ind in ('Y','N'));
+comment on column ks_sessions.presented_anything_ind is 'Whether the presenter has ever done a live presentation(s), anywhere for anything.';
+comment on column ks_sessions.presented_anything_where is 'Where the presenter has done live presentation (of any kind)';
 
+comment on column ks_sessions.presented_before_ind is 'Whether the session has been presented before';
+comment on column ks_sessions.presented_before_where is 'Where the presentaton been done before';
 
-@../install/ks_user_event_track_roles.sql
 
+-- #20
+alter table ks_session_votes add decline_vote_flag varchar2(1);
+comment on column ks_session_votes.decline_vote_flag is 'Used when a user abstains form voting on a session.';
 
 
--- =============================================================================
-@../conversion/move_owners_to_roles.sql
+-- #3
+@../install/ks_event_admins.sql
+@../install/ks_email_templates.sql
 
--- drop track owner from the track instance
-alter table ks_event_tracks drop column owner;
 
 
 
+-- *** Views ***
 
--- =============================================================================
--- add voting date fields to events and event-tracks
-alter table ks_events add (
-    blind_vote_begin_date  date
-  , blind_vote_end_date    date
-  , committee_vote_begin_date date
-  , committee_vote_end_date   date
-)
-/
-comment on column ks_events.blind_vote_begin_date is 'Begin date of blind voting';
-comment on column ks_events.blind_vote_end_date is 'End date of blind voting';
-comment on column ks_events.committee_vote_begin_date is 'Begin date of committee voting';
-comment on column ks_events.committee_vote_end_date is 'End date of committee voting';
-
-alter table ks_event_tracks add (
-    blind_vote_begin_date  date
-  , blind_vote_end_date    date
-  , committee_vote_begin_date date
-  , committee_vote_end_date   date
-)
-/
-
-comment on column ks_event_tracks.blind_vote_begin_date is 'Track Override for Begin date of blind voting';
-comment on column ks_event_tracks.blind_vote_end_date is 'Track Override for End date of blind voting';
-comment on column ks_event_tracks.committee_vote_begin_date is 'Track Override for Begin date of committee voting';
-comment on column ks_event_tracks.committee_vote_end_date is 'Track Override for End date of committee voting';
-
-
--- =============================================================================
-alter table ks_event_tracks add (
-    blind_vote_help        varchar2(4000)
-  , committee_vote_help    varchar2(4000)
-)
-/
-comment on column ks_event_tracks.blind_vote_help is 'Help for blind voters';
-comment on column ks_event_tracks.committee_vote_help is 'Help for committee voters';
-
-
-
-
--- =============================================================================
-alter table ks_users add external_sys_ref varchar2(20);
-comment on column ks_users.external_sys_ref is 'Unique "ID" for external system.';
-
-
-
-
-
--- =============================================================================
-
---Update ks_sessions to allow for scoring, and other input from the user's uploading of an excel file.
-alter table ks_sessions add presenter_email varchar2(500);
-alter table ks_sessions add session_summary varchar2(4000);
-alter table ks_sessions add session_abstract clob;
-alter table ks_sessions add target_audience varchar2(60);
-alter table ks_sessions add presented_before_ind varchar2(1) default on null 'N';
-alter table ks_sessions add presented_before_where varchar2(4000);
-alter table ks_sessions add technology_product varchar2(500);
-alter table ks_sessions add ace_level varchar2(30);
-alter table ks_sessions add video_link varchar2(4000);
-alter table ks_sessions add contains_demo_ind varchar2(1) default on null 'N';
-alter table ks_sessions add webinar_willing_ind varchar2(1) default on null 'N';
-alter table ks_sessions add presenter_biography clob;
-alter table ks_sessions add external_sys_ref varchar2(20);
-alter table ks_sessions add presenter_user_id varchar2(20);
-alter table ks_sessions add co_presenter_user_id varchar2(20);
-
-alter table ks_sessions add constraint sessions_ck_presb4_is_yn check (presented_before_ind in ('Y', 'N'));
-alter table ks_sessions add constraint sessions_ck_demo_is_yn check (contains_demo_ind in ('Y', 'N'));
-alter table ks_sessions add constraint sessions_ck_web_will_is_yn check (webinar_willing_ind in ('Y', 'N'));
-
-comment on column ks_sessions.presenter_email is 'The email address to contact the presenter.';
-comment on column ks_sessions.session_summary is 'Summary/Short description. This is cut off at 4000 characters since anything submitted that''s longer is not short.';
-comment on column ks_sessions.session_abstract is 'The full abstract of the session. This is a clob b/c it can be > 4000 characters';
-comment on column ks_sessions.target_audience is 'The audience the session is intended for. Can be multiple audiences.';
-comment on column ks_sessions.presented_before_ind is 'Whether the presenter has presented this session before.';
-comment on column ks_sessions.presented_before_where is 'Where the presenter has presented this session before. ';
-comment on column ks_sessions.technology_product is 'Technologies or products discussed in the presentation.';
-comment on column ks_sessions.ace_level is 'Whether the presenter is part of an ace program. The common values are No, Oracle Ace, Oracle Ace Director, Oracle Ace Associate.';
-comment on column ks_sessions.video_link is 'Link to a video for furth info about the session.';
-comment on column ks_sessions.contains_demo_ind is 'Whether a demo is included. Correct values or Y or N.';
-comment on column ks_sessions.webinar_willing_ind is 'Whether the presenter is willing to do the session as a webinar. Correct values are Y or N.';
-comment on column ks_sessions.presenter_biography is 'The biography of this session''s presenter.';
-comment on column ks_sessions.external_sys_ref is 'Unique "ID" for external system.';
-comment on column ks_sessions.presenter_user_id is 'External System User ID of the presenter';
-comment on column ks_sessions.co_presenter_user_id is 'External System User ID of the co-presenter';
-
-
-alter table ks_sessions modify tags varchar2(1000);
-
-
-
-
-
--- =============================================================================
--- adding event-types
-@../install/ks_event_types.sql
-@../conversion/seed_ks_event_types.sql
-
-alter table ks_events add
-  event_type varchar2(20);
-
-comment on column ks_events.event_type is 'type of event';
-
-update ks_events
-  set event_type = 'KSCOPE';
-
--- add foreign key constraint from events to event-types
-alter table ks_events add constraint ks_events_event_type_fk foreign key (event_type) references ks_event_types (code);
-alter table ks_events modify event_type not null;
-
-
-
-
-
--- =============================================================================
-
--- modifications for voting by community and committee
-alter table ks_session_votes add (
-    vote_type varchar2(10)  -- COMMITTEE | BLIND
-  , username  varchar2(60)
-)
-/
-
-comment on column ks_session_votes.vote_type is 'Identifies the type of vote recorded: COMMITTEE or BLIND';
-comment on column ks_session_votes.username is 'User that casted the vote.';
-
-
-alter table ks_session_votes modify vote_type not null;
-alter table ks_session_votes modify username not null;
-
-
---Now Add users that are existent in ks_session_voter but not in ks_users
-merge into ks_users u
-  using (select distinct username, voter from ks_session_votes) v
-  on (u.username = v.username)
-  when not matched
-    then
-      insert (username, first_name, active_ind, admin_ind) values (v.username, v.voter, 'Y', 'N')
-/
-
---Finally drop the voter column
-alter table ks_session_votes drop column voter
-/
-
--- one vote per user per session
-create unique index ks_session_votes_u01 on ks_session_votes(username, session_id);
-
-
--- =============================================================================
--- Seed ks_parameters
-delete from ks_parameters where name_key = 'DEFAULT_BLIND_VOTING_HELP';
-insert into ks_parameters(category, name_key, value, description)
-                   values('Messages',
-                          'DEFAULT_BLIND_VOTING_HELP'
-                        , 'Vote on all sessions to the best of your ability.<br>'
-                       || '1 is the lowest rating and 5 is the highest. Use 1 or 2 for sessions you have no interest in.'
-                        , 'The default help message for blind voters viewed while voting on a track.');
-
-delete from ks_parameters where name_key = 'DEFAULT_COMMITTEE_VOTING_HELP';
-insert into ks_parameters(category, name_key, value, description)
-                   values('Messages'
-                        , 'DEFAULT_COMMITTEE_VOTING_HELP'
-                        , 'Vote on all sessions to the best of your ability.<br>'
-                       || '1 is the lowest rating and 5 is the highest. Use 1 or 2 for sessions you have no interest in.'
-                        , 'The default help message for blind voters viewed while voting on a track.');
-
-delete from ks_parameters where name_key = 'REVIEW_PRIVACY_NOTICE';
-insert into ks_parameters(category, name_key, value, description)
-                   values('Messages',
-                          'REVIEW_PRIVACY_NOTICE',
-                          'Remember, all the votes and comments are private information and can only be discussed with other members of the committee.',
-                          'This message is viewed once by anyone using the review app.');
-
-delete from ks_parameters where name_key = 'VOTING_PRIVACY_NOTICE';
-insert into ks_parameters(category, name_key, value, description)
-                   values('Messages',
-                          'VOTING_PRIVACY_NOTICE',
-                          'Remember, all the votes and comments are private information and can only be discussed with the content committee.',
-                          'This message is viewed once by anyone using the voting app.');
-
-
--- 
-delete from ks_parameters where name_key = 'ANONYMIZE_EXTRA_TOKENS';
-insert into ks_parameters(category, name_key, value, description)
-                   values('Sessions'
-                        , 'ANONYMIZE_EXTRA_TOKENS'
-                        , ''
-                        , 'Case-insensitive, space-delimited list of words to anonymize. E.g. "Word1 Word2 Word3"');
-
--- 
-delete from ks_parameters where name_key in ('EMAIL_FROM_ADDRESS', 'FEEDBACK_EMAIL', 'EMAIL_OVERRIDE', 'EMAIL_PREFIX');
-insert into ks_parameters(category, name_key, value, description) values ('SYSTEM', 'EMAIL_FROM_ADDRESS', 'info@noname.com', 'Email addresses used to send emails.');
-insert into ks_parameters(category, name_key, value, description) values ('SYSTEM', 'FEEDBACK_EMAIL', 'info@noname.com', 'Email addresses that will receive system feedback.');
-insert into ks_parameters(category, name_key, value, description) values ('SYSTEM', 'EMAIL_OVERRIDE', '', 'Comma delimited list of emails that will be used to override all emails from the system.  For use during dev and test.');
-insert into ks_parameters(category, name_key, value, description) values ('SYSTEM', 'EMAIL_PREFIX', '[VOTEAPP] ', 'Prefix used when sending emails');
-
-delete from ks_parameters where name_key in ('VOTING_APP_ID');
-insert into ks_parameters(category, name_key, value, description) values ('SYSTEM', 'VOTING_APP_ID', '120124', 'ID of voting app');
-
--- =============================================================================
-
-@../conversion/seed_constraint_lookup.sql
-
-
-
-
--- =============================================================================
-
+-- #3
 @../views/ks_users_v.sql
+
 @../views/ks_events_tracks_v.sql
+@../views/ks_events_sec_v.sql
+@../views/ks_events_allowed_v.sql
+
+-- #16
+@../views/ks_session_load_coll_v.sql
+
+-- #32 
+@../views/ks_sessions_v.sql
 
 
 
 
--- =============================================================================
-PRO PACKAGES
-PRO .Specs
---install .xlsx reader
-PRO .package_read_xlsx_clob
-@../plsql/package_read_xlsx_clob.pks
-PRO .ks_log
-@../plsql/ks_log.pls
-PRO .ks_user_dml
-@../plsql/ks_user_dml.pls
-PRO .ks_util
-@../plsql/ks_util.pls
-PRO .ks_session_load_api
-@../plsql/ks_session_load_api.pls
-PRO .ks_users_api
-@../plsql/ks_users_api.pls
-PRO .ks_session_api
+
+-- *** Objects ***
+
+-- Added ks_log calls
+@../plsql/ks_tags_api.plb
+
+-- #4
+@../plsql/ks_error_handler.plb
+
+-- #1, #22, #20, #32
 @../plsql/ks_session_api.pls
-PRO .ks_user_event_track_roles_dml
-@../plsql/ks_user_event_track_roles_dml.pls
-
-PRO .Body
-PRO .ks_log
-@../plsql/ks_log.plb
-PRO ks_user_dml
-@../plsql/ks_user_dml.plb
-PRO package_read_xlsx_clob
-@../plsql/package_read_xlsx_clob.pkb
-PRO ks_util
-@../plsql/ks_util.plb
-PRO ks_session_load_api
-@../plsql/ks_session_load_api.plb
-PRO ks_users_api
-@../plsql/ks_users_api.plb
-PRO .ks_session_api
 @../plsql/ks_session_api.plb
-PRO ks_user_event_track_roles_dml
-@../plsql/ks_user_event_track_roles_dml.plb
 
+-- #16
+@../plsql/ks_session_load_api.pls
+@../plsql/ks_session_load_api.plb
+
+-- #6, #13
+@../plsql/ks_email_api.pls
+@../plsql/ks_email_api.plb
+@../plsql/ks_notification_api.pls
+@../plsql/ks_notification_api.plb
+
+-- #13
+@../plsql/ks_sec.pls
+@../plsql/ks_sec.plb
+
+
+
+
+-- *** DML ***
+insert into constraint_lookup (constraint_name,message) values ('KS_USERNAME_U','User already exists.');
+
+delete from ks_parameters where name_key in ('ADMIN_APP_ID');
+insert into ks_parameters(category, name_key, value, description) values ('SYSTEM', 'ADMIN_APP_ID', '83791', 'ID of Admin app');
+
+delete from ks_parameters where name_key in ('SERVER_URL');
+insert into ks_parameters(category, name_key, value, description) values ('SYSTEM', 'SERVER_URL', 'https://apex.oracle.com/pls/apex/f?p=', 'Server URL');
+
+delete from ks_parameters where name_key in ('LOAD_NOTIFICATION_TEMPLATE');
+insert into ks_parameters(category, name_key, value, description) values ('Notifications', 'LOAD_NOTIFICATION_TEMPLATE', 'SESSION_LOAD', 'Name of email template for load notifications');
+
+
+insert into ks_parameters (category,name_key,value,description) values ('Notifications','RESET_PASSWORD_DONE_NOTIFICATION_TEMPLATE','RESET_PASSWORD_DONE_NOTIFICATION','Name of email template for when a reset password is executed');
+insert into ks_parameters (category,name_key,value,description) values ('Notifications','RESET_PASSWORD_REQUEST_NOTIFICATION_TEMPLATE','RESET_PASSWORD_REQUEST_NOTIFICATION','Name of email template for reset password request notifications');
+
+
+update ks_roles
+   set name = 'Public Voter'
+ where role_type = 'VOTING'
+   and code = 'BLIND';
+
+
+-- #16
+-- New columns presented_anything_ind, presented_anything_where
+-- fix order
+@../conversion/seed_ks_load_mapping.sql
+@../conversion/seed_ks_email_templates.sql
+
+
+-- DO NOT TOUCH/UPDATE BELOW THIS LINE
+
+
+PRO Recompiling objects
+exec dbms_utility.compile_schema(schema => user, compile_all => false);